requirements-ai.txt
# AI-Accelerated Asset Pipeline Requirements
# Install with: pip install -r requirements-ai.txt

# Core image processing
Pillow>=10.0.0

# AI APIs (NEW google-genai package, not google-generativeai)
google-genai>=0.1.0

# ML-based background removal (optional but recommended)
# rembg>=2.0.0

# Color clustering for palette optimization
scikit-learn>=1.3.0

# Environment variable management
python-dotenv>=1.0.0

# Optional: Fast inference with Groq
# groq>=0.4.0

# Development/testing
pytest>=7.4.0
